use intptr_t to cast from ptr to int for masking low bits
git-svn-id: http://skia.googlecode.com/svn/trunk@4800 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/opts/SkBlitRow_opts_SSE2.cpp
@@ -900,7 +900,7 @@ void S32A_D565_Blend_SSE2(uint16_t* SK_RESTRICT dst,
 void S32A_D565_Opaque_SSE2(uint16_t* SK_RESTRICT dst,
                                const SkPMColor* SK_RESTRICT src, int count,
                                U8CPU alpha, int /*x*/, int /*y*/) {
-    while(((int)src & 0xf) && (count-- > 0)){
+    while(((intptr_t)src & 0xf) && (count-- > 0)){
                 *dst = SkSrcOver32To16(*src, *dst);
         src++;
         dst++;
